MIDI-mapping and virtual keyboard
The table showing the MIDI note and virtual keyboard mappings for the GMRockKit was wrong for some instruments. Also, it suggested there is a one to one mapping of instrument number, MIDI note, and virtual key - which is not the key. I removed the instrument number column and shuffled the others to emphasis that the table is just about highlighting settings of the GMRockKit.

The image showing two octaves of a keyboard and mapping it to PC keys was removed. It suggested a mapping of virtual keys to specific MIDI notes (or instruments assigned to them) which is not what Hydrogen does. Instead, it maps the virtual keys to the instrument position. This is visualized in a new image showing a PC keyboard with the corresponding instrument numbers added.

The sections covering the MIDI note input and the virtual keyboard have been fixed