fix cmake installs
Loads of unrequired files and the whole folder hierarchy (most empty) were installed but some key resources, like the CSS and JS file for both the manual and the tutorials, were omitted.

I also narrowed the installed manuals to just the English version and only the images it requires. For the tutorial all translations will be installed as well.

Fixes #45