relicense repo as GPLv2+
This repo was once split off the main hydrogen one. But, unfortunately, the copyright file was missed when disentangling the two.

With this commit we relicense the Hydrogen documentation under the same license as the main Hydrogen project. All (non-member) contributors gave their consent and the active members of the Hydrogen development team give their consent on behalf of the former/inactive members to whom we have no means of contact anymore.

For a detailed discussion see https://github.com/hydrogen-music/documentation/issues/73.

Addresses #73