move outdated translations
into a dedicated folder. With the fuzzy matching active they do way more harm than good. Quite a number of fixes to the individual translations are necessary when changing content in manual.docbook and as of now all translations have fallen into despair and diverged to such an extend from the English version that we can not ship them anymore