cmake: first draft of CMake port
streamlining installation process.

Build artifacts (HTML files) during build. This is done by calling `make` during the build step. But `make` is still hard coded, the requirements of the toolchains are not checked yet, there is an error, and the output is not nicely formatted. But in general it works and the artifacts are created