manual: remove JACK recommendation
lately JACK caused quite a number of problems and is getting more and more unstable (especially on Windows but also our macOS has to be rerun from time to time since `jackd` segfaulted). For Linux users who want to combine Hydrogen with other application this is still the go-to option. But all other users, especially on other platforms, might make better use of other drivers